Ignore image file header, geometrical transformations only will be read/write to metadatas, except header_* programs Changing select_img from int to size_t, also Ndim, now default value will be ALL_IMAGES = 0, index logic changes. Changed readdata(boolean) by datamode(enum) ... Some re-organization on _read and readApplyGeo... And more.....
git-svn-id: file:///home/roberto/Git/svn@7407 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 235036006c244f34bc8b2c3bf5a3dd43e04a0887

+++ trunk/xmipp/applications/programs/header_assign/header_assign_main.cpp
@@ -133,10 +133,10 @@ protected:
                 break;
             }
         }
-        int no;
+        size_t no;
         FileName stackFn;
         fnImgOut.decompose(no, stackFn);
-        img.write(stackFn,no,no!=-1,WRITE_REPLACE);
+        img.write(stackFn,no,no!=-1, WRITE_REPLACE);
     }
 }
 ;// end of class ProgHeaderAssign
